Server/requirements.txt
Django==3.2.12
djangorestframework==3.13.1
drf-nested-routers==0.93.4
requests